requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-environ==0.11.2
psycopg2-binary==2.9.9
requests==2.32.3
geopy==2.4.1
pandas==2.1.4
gunicorn==22.0.0
pytest==8.2.2
pytest-django==4.8.0